format error fixes

+++ src/asn.c
@@ -264,8 +264,8 @@ asHandleReply(void *data, char *buf, ssize_t size)
     }
     asState->seen = asState->offset + size;
     asState->offset += (s - buf);
-    debug(53, 3) ("asState->seen = %d, asState->offset = %d\n",
-	asState->seen, asState->offset);
+    debug(53, 3) ("asState->seen = %ld, asState->offset = %ld\n",
+	(long int)asState->seen, (long int)asState->offset);
     if (e->store_status == STORE_PENDING) {
 	debug(53, 3) ("asHandleReply: store_status == STORE_PENDING: %s\n", storeUrl(e));
 	storeClientCopy(asState->sc,
